Backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sqlite-utils
PyJWT
requests
reportlab